+++ queue_job/readme/HISTORY.rst
@@ -13,6 +13,8 @@ Next
 
 * [ADD] Run jobrunner as a worker process instead of a thread in the main
   process (when running with --workers > 0)
+* [REF] ``@job`` and ``@related_action`` deprecated, any method can be delayed,
+  and configured using ``queue.job.function`` records
 
 
 13.0.1.2.0 (2020-03-10)